backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
alembic
psycopg2-binary
python-multipart
python-dotenv
google-genai
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
